requirements.txt
cffi==1.16.0
cryptography==41.0.7
mysql-connector-python==8.2.0
mysqlclient==2.2.1
protobuf==4.21.12
pycparser==2.21
pycryptodome==3.19.1
pyope==0.2.2
python-dotenv==1.0.0
six==1.16.0
sqlglot==20.7.1
tabulate==0.9.0
